reserve/requirements.txt
pygame>=2.5.0
numpy>=1.24.0
pydub>=0.25.1
matplotlib>=3.7.0
pillow>=10.0.0
PyQt5>=5.15.0
scipy>=1.10.0
